feat: 完善 llm_runner 批量推理示例
- LLMEngine 支持 generate_stream 聊天模板与 close 清理
- ModelWorker 实现真实推理/采样，支持并行进程
- 补充 batch2 多模型示例配置（样例路径）

+++ src/qwen3_from_scratch/models/attn.py
@@ -371,7 +371,7 @@ class VarLenPagedAttn(MyAttn):
 
                         factor = torch.exp(max_val - curr_max)
                         dominator = dominator * factor + curr_dominator
-                        curr_output = factor * curr_output + attn_score @ sub_v
+                        curr_output = factor * curr_output + attn_score.to(sub_v.dtype) @ sub_v
 
                         max_val = curr_max
                     output[m:m + curr_m_span, h] = curr_output / dominator
